requirements.txt
# ── Runtime (producer + scripts) ─────────────────────────────────────────────
kafka-python==2.0.2
psycopg2-binary==2.9.9
requests==2.31.0
fastavro==1.9.4
python-dotenv==1.0.1

# ── Airflow (DAGs — même version que Docker) ──────────────────────────────────
apache-airflow==2.10.4

# ── Dev / type-checking ───────────────────────────────────────────────────────
pyrefly
